requirements.txt
streamlit
transformers
torch
Pillow
accelerate
torchvision
numpy